\name{vworld.map-package}
\alias{vworld.map}
\alias{vworld.map-package}
\docType{package}
\title{vworld.map: VWorld StaticMap API 2.0 client for R}
\description{
Build VWorld StaticMap (GetMap) request URLs or download static maps from R,
with markers, routes, overlay layers, custom CRS, and ggplot2 / leaflet
integration.
}
\seealso{
\code{\link{vworld_static_map_url}},
\code{\link{vworld_static_map}},
\code{\link{vworld_marker}},
\code{\link{vworld_route}},
\code{\link{vworld_ggmap_layer}},
\code{\link{vworld_leaflet_image}}
}
