\name{VWORLD_BASEMAPS}
\alias{VWORLD_BASEMAPS}
\alias{VWORLD_CRS}
\title{VWorld constants: basemap codes and supported CRS}
\format{
\code{VWORLD_BASEMAPS}: a character vector of allowed \code{basemap} codes.
\code{VWORLD_CRS}: a named character vector of EPSG codes.
}
\description{
Static lookup tables used by \code{\link{vworld_static_map_url}} to validate
the \code{basemap} and \code{crs} arguments.
}
